community/hashnode-blog-translator/pyproject.toml
[project]
name = "hashnode-blog-translator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=6.4.0",
    "httpx>=0.28.1",
    "lingodotdev>=1.3.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
]
